refactor: 로고만 있는 헤더를 컴포넌트로 분리
- 기존 Header 컴포넌트는 MainHeader로 이름 수정
- 로그인/회원가입의 헤더는 LogoHeader라는 이름의 컴포넌트로 분리
- 둘 다 헤더이기 때문에 둘 다 같은 레벨인 컴포넌트로 만듦
