+++ components/Layout.tsx
@@ -1,9 +1,8 @@
-import Image from 'next/image';
-import Link from 'next/link';
 import { useRouter } from 'next/router';
 
 import Footer from './Footer';
-import Header from './Header';
+import LogoHeader from './LogoHeader';
+import MainHeader from './MainHeader';
 
 type Props = {
   children: React.ReactNode;
@@ -15,7 +14,7 @@ export default function Layout({ children }: Props) {
   if (asPath === '/') {
     return (
       <>
-        <Header />
+        <MainHeader />
         <main className={`mt-[70px]`}>{children}</main>
         <Footer />
       </>
@@ -25,18 +24,7 @@ export default function Layout({ children }: Props) {
   if (asPath === '/signin' || asPath === '/signup') {
     return (
       <>
-        <header className="flex justify-center pt-6 tablet:pt-12 desktop:pt-16">
-          <Link href="/">
-            <Image
-              width={198}
-              height={66}
-              className="tablet:h-[132px] tablet:w-[396px]"
-              src="/images/logo_big.svg"
-              alt="판다마켓 로고"
-              priority
-            />
-          </Link>
-        </header>
+        <LogoHeader />
         <main className="m-auto px-4 pb-20 pt-6 tablet:px-12 tablet:pb-24 desktop:px-16 desktop:pb-16">
           {children}
         </main>
@@ -46,7 +34,7 @@ export default function Layout({ children }: Props) {
 
   return (
     <>
-      <Header />
+      <MainHeader />
       <main
         className={`m-auto mt-[70px] px-4 pb-24 pt-8 tablet:px-6 desktop:max-w-[1200px] desktop:px-0`}
       >
